+++ CHANGELOG.md
@@ -2,6 +2,33 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.18.0] - 2025-12-05
+
+### 🐛 Bug Fixes
+
+- Error message for invalid config files (#223)
+- [**breaking**] Reverse `env_files` precedence (#224)
+
+### ⚙️ Miscellaneous Tasks
+
+- Add license
+
+## [0.17.0] - 2025-12-04
+
+### 🚀 Features
+
+- Add `watch` as a built-in var (#222)
+
+### 📚 Documentation
+
+- Env precedence
+
+## [0.16.1] - 2025-12-04
+
+### 🐛 Bug Fixes
+
+- CLI vars evaluation order (#221)
+
 ## [0.16.0] - 2025-12-01
 
 ### 🚀 Features